requirements.txt
Flask==1.1.2
google-cloud-storage